Use the maven-bundle-plugin "manifest" goal to generate the MANIFEST.MF file and configure the maven-jar-plugin to use the generated MANIFEST.MF.
The main advantage of this is that *normal* manifest entries come first and retain their order, with the OSGi entries being appended after. It also means that the component poms don't need to specify <packaging>bundle</packaging>. Ovrriding the "commons.manifestfile" property with an alternative (or blank) manifest file can be used to omit OSGi manifest entries being added.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/commons-parent/trunk@633514 13f79535-47bb-0310-9956-ffa450edef68